+++ LoggingModule/LoggingModule.cs
@@ -15,160 +15,52 @@ namespace SyslogLogging
     /// Syslog, console, and file logging module.
     /// </summary>
     public class LoggingModule : IDisposable
-    { 
+    {
         #region Public-Members
 
         /// <summary>
-        /// Server IP address.
+        /// Logging settings.
         /// </summary>
-        public string ServerIp
+        public LoggingSettings Settings
         {
             get
             {
-                return _ServerIp;
+                return _Settings;
             }
             set
             {
-                if (String.IsNullOrEmpty(value)) throw new ArgumentNullException(nameof(ServerIp));
-
-                IPAddress addr = null;
-                if (!IPAddress.TryParse(value, out addr)) throw new ArgumentException("Use an IP address instead of a hostname.");
-
-                _ServerIp = value;
+                if (value == null) _Settings = new LoggingSettings();
+                else _Settings = value;
             }
         }
 
         /// <summary>
-        /// UDP port on which the syslog server is listening.
+        /// List of syslog servers.
         /// </summary>
-        public int ServerPort
+        public List<SyslogServer> Servers
         {
             get
             {
-                return _ServerPort;
+                return _Servers;
             }
             set
             {
-                if (value < 0 || value > 65535) throw new ArgumentException("Port must be in the range 0-65535.");
-                _ServerPort = value;
+                if (value == null) _Servers = new List<SyslogServer>();
+                else _Servers = value;
             }
         }
-
-        /// <summary>
-        /// Enable or disable console logging.
-        /// </summary>
-        public bool ConsoleEnable
-        {
-            get
-            {
-                return _ConsoleEnable;
-            }
-            set
-            {
-                if (value && !ConsoleExists())
-                {
-                    _ConsoleEnable = false;
-                }
-                else
-                {
-                    _ConsoleEnable = value;
-                }
-            }
-        }
-
-        /// <summary>
-        /// Minimum severity required to send a message.
-        /// </summary>
-        public Severity MinimumSeverity = Severity.Debug;
-
-        /// <summary>
-        /// Enable or disable async logging.
-        /// </summary>
-        public bool AsyncLogging = false;
-
-        /// <summary>
-        /// Include the UTC timestamp in the message.
-        /// </summary>
-        public bool IncludeUtcTimestamp = true;
-
-        /// <summary>
-        /// Include the severity in the message.
-        /// </summary>s
-        public bool IncludeSeverity = true;
-
-        /// <summary>
-        /// Include the local hostname in the message.
-        /// </summary>
-        public bool IncludeHostname = false;
-
-        /// <summary>
-        /// Include the local thread ID in the message.
-        /// </summary>
-        public bool IncludeThreadId = false;
-
-        /// <summary>
-        /// Indent outgoing messages based on stack depth.
-        /// </summary>
-        public bool IndentByStackSize = false;
-
-        /// <summary>
-        /// Enable or disable use of color for console messages.
-        /// </summary>
-        public bool EnableColors = true;
-
-        /// <summary>
-        /// Colors to use for console messages based on message severity.
-        /// </summary>
-        public ColorSchema Colors = new ColorSchema();
-
-        /// <summary>
-        /// Enable or disable logging to a file.
-        /// Disabled: file logging will not be used.
-        /// SingleLogFile: all messages will be appended to a single file.
-        /// FileWithDate: all messages will be appended to a file, where the name of the file is the supplied filename followed by '.yyyyMMdd'.
-        /// </summary>
-        public FileLoggingMode FileLogging = FileLoggingMode.Disabled;
-
-        /// <summary>
-        /// The file to which log messages should be appended.
-        /// </summary>
-        public string LogFilename = null;
-
-        /// <summary>
-        /// The severity level to use when logging exceptions through the .Exception() method.  
-        /// </summary>
-        public Severity ExceptionSeverity = Severity.Alert;
-
-        /// <summary>
-        /// Maximum message length.  Must be greater than or equal to 32.  Default is 1024.
-        /// </summary>
-        public int MaxMessageLength
-        {
-            get
-            {
-                return _MaxMessageLength;
-            }
-            set
-            {
-                if (value < 32) throw new ArgumentException("Maximum message length must be at least 32.");
-                _MaxMessageLength = value;
-            }
-        }
-
+         
         #endregion
 
         #region Private-Members
 
         private bool _Disposed = false;
-        private string _ServerIp = "127.0.0.1";
-        private int _ServerPort = 514; 
-        private UdpClient _UDP = null;
-        private string _Hostname = null;
-        private object _SendLock = new object();
-        private int _BaseDepth = 0;
-        private bool _ConsoleEnable = true;
-        private int _MaxMessageLength = 1024;
+        private LoggingSettings _Settings = new LoggingSettings();
+        private List<SyslogServer> _Servers = new List<SyslogServer> { new SyslogServer("127.0.0.1", 514) };
         private readonly object _FileLock = new object();
+        private string _Hostname = Dns.GetHostName();
+        private CancellationTokenSource _TokenSource = new CancellationTokenSource();
+        private CancellationToken _Token;
 
         #endregion
 
@@ -179,14 +71,7 @@ namespace SyslogLogging
         /// </summary>
         public LoggingModule()
         {
-            ServerIp = "127.0.0.1";
-            ServerPort = 514;
-
-            _UDP = new UdpClient(ServerIp, ServerPort);
-            _Hostname = Dns.GetHostName();
-
-            StackTrace st = new StackTrace();
-            _BaseDepth = st.FrameCount - 1;
+            _Token = _TokenSource.Token;
         }
 
         /// <summary>
@@ -194,46 +79,33 @@ namespace SyslogLogging
         /// </summary>
         /// <param name="serverIp">Server IP address.</param>
         /// <param name="serverPort">Server port number.</param>
+        /// <param name="enableConsole">Enable or disable console logging.</param>
         public LoggingModule(
             string serverIp,
-            int serverPort)
+            int serverPort,
+            bool enableConsole = true)
         {
-            ServerIp = serverIp;
-            ServerPort = serverPort;
-
-            _UDP = new UdpClient(ServerIp, ServerPort);
-            _Hostname = Dns.GetHostName();
-
-            StackTrace st = new StackTrace();
-            _BaseDepth = st.FrameCount - 1; 
+            SyslogServer server = new SyslogServer(serverIp, serverPort);
+            _Servers.Add(server);
+            _Settings.EnableConsole = enableConsole;
+            _Token = _TokenSource.Token;
         }
 
         /// <summary>
-        /// Instantiate the object to enable either file logging or console logging.
+        /// Instantiate the object using a series of servers.
         /// </summary>
-        /// <param name="filename">Filename.</param>
+        /// <param name="servers">Servers.</param>
         /// <param name="enableConsole">Enable or disable console logging.</param>
         public LoggingModule(
-            string filename,
-            bool enableConsole)
+            List<SyslogServer> servers,
+            bool enableConsole = true)
         {
-            if (String.IsNullOrEmpty(filename) && !enableConsole) throw new ArgumentException("Either a filename must be specified or console logging must be enabled.");
+            if (servers == null) throw new ArgumentNullException(nameof(servers));
+            if (servers.Count < 1) throw new ArgumentException("At least one server must be specified.");
 
-            ServerIp = "127.0.0.1";
-            ServerPort = 514;
-            _UDP = null;
-            _Hostname = null;
-
-            StackTrace st = new StackTrace();
-            _BaseDepth = st.FrameCount - 1;
-
-            if (!String.IsNullOrEmpty(filename))
-            {
-                LogFilename = filename;
-                FileLogging = FileLoggingMode.SingleLogFile;
-            }
-
-            ConsoleEnable = enableConsole;
+            _Servers = servers;
+            _Settings.EnableConsole = enableConsole;
+            _Token = _TokenSource.Token;
         }
 
         /// <summary>
@@ -244,75 +116,21 @@ namespace SyslogLogging
         /// <param name="enableConsole">Enable or disable console logging.</param>
         public LoggingModule(
             string filename,
-            FileLoggingMode fileLoggingMode,
-            bool enableConsole)
+            FileLoggingMode fileLoggingMode = FileLoggingMode.SingleLogFile,
+            bool enableConsole = true)
         {
-            if (String.IsNullOrEmpty(filename)) throw new ArgumentNullException(nameof(filename));
+            if (String.IsNullOrEmpty(filename) && !enableConsole) throw new ArgumentException("Either a filename must be specified or console logging must be enabled.");
 
-            ServerIp = "127.0.0.1";
-            ServerPort = 514;
-            _UDP = null;
-            _Hostname = null;
-
-            StackTrace st = new StackTrace();
-            _BaseDepth = st.FrameCount - 1;
-
-            LogFilename = filename;
-            FileLogging = fileLoggingMode;
-
-            ConsoleEnable = enableConsole;
+            _Settings.FileLogging = fileLoggingMode;
+            _Settings.LogFilename = filename;
+            _Settings.EnableConsole = enableConsole;
+            _Token = _TokenSource.Token;
         }
-
-        /// <summary>
-        /// Instantiate the object.
-        /// </summary>
-        /// <param name="serverIp">Server IP address.</param>
-        /// <param name="serverPort">Server port number.</param>
-        /// <param name="consoleEnable">Enable or disable console logging.</param>
-        /// <param name="minimumSeverity">Minimum severity required to send a message.</param>
-        /// <param name="asyncLogging">Enable or disable async logging.</param>
-        /// <param name="includeUtcTimestamp">Include the UTC timestamp in the message.</param>
-        /// <param name="includeSeverity">Include the severity in the message.</param>
-        /// <param name="includeHostname">Include the local hostname in the message.</param>
-        /// <param name="includeThreadId">Include the local thread ID in the message.</param>
-        /// <param name="indentByStackSize">Indent outgoing messages based on stack depth.</param>
-        public LoggingModule(
-            string serverIp,
-            int serverPort,
-            bool consoleEnable,
-            Severity minimumSeverity,
-            bool asyncLogging,
-            bool includeUtcTimestamp,
-            bool includeSeverity,
-            bool includeHostname,
-            bool includeThreadId,
-            bool indentByStackSize)
-        {
-            if (String.IsNullOrEmpty(serverIp)) throw new ArgumentNullException(nameof(serverIp));
-            if (serverPort < 0 && serverPort > 65535) throw new ArgumentException("Server port must in the range 0-65535.");
-
-            ServerIp = serverIp;
-            ServerPort = serverPort;
-            ConsoleEnable = consoleEnable;
-            MinimumSeverity = minimumSeverity;
-            AsyncLogging = asyncLogging;
-            IncludeUtcTimestamp = includeUtcTimestamp;
-            IncludeSeverity = includeSeverity;
-            IncludeHostname = includeHostname;
-            IncludeThreadId = includeThreadId;
-            IndentByStackSize = indentByStackSize;
-               
-            _UDP = new UdpClient(ServerIp, ServerPort); 
-            _Hostname = Dns.GetHostName();
-
-            StackTrace st = new StackTrace();
-            _BaseDepth = st.FrameCount - 1;
-        }
-
+       
         #endregion
 
         #region Public-Methods
-         
+
         /// <summary>
         /// Tear down the client and dispose of background workers.
         /// </summary>
@@ -398,7 +216,7 @@ namespace SyslogLogging
         /// <param name="module">Module name (user-specified).</param>
         /// <param name="method">Method name (user-specified).</param>
         /// <param name="e">Exception.</param>
-        public virtual void Exception(string module, string method, Exception e)
+        public virtual void Exception(Exception e, string module = null, string method = null)
         {
             if (e == null) throw new ArgumentNullException(nameof(e));
             var st = new StackTrace(e, true);
@@ -409,8 +227,8 @@ namespace SyslogLogging
             string message =
                 Environment.NewLine +
                 "--- Exception details ---" + Environment.NewLine +
-                "  Module     : " + module + Environment.NewLine +
-                "  Method     : " + method + Environment.NewLine +
+                (!String.IsNullOrEmpty(module) ? "  Module     : " + module + Environment.NewLine : "") +
+                (!String.IsNullOrEmpty(method) ? "  Method     : " + method + Environment.NewLine : "") +
                 "  Type       : " + e.GetType().ToString() + Environment.NewLine;
 
             if (e.Data != null && e.Data.Count > 0)
@@ -457,13 +275,12 @@ namespace SyslogLogging
                 "  Message    : " + e.Message + Environment.NewLine +
                 "  Source     : " + e.Source + Environment.NewLine +
                 "  StackTrace : " + e.StackTrace + Environment.NewLine +
-                "  Stack      : " + StackToString() + Environment.NewLine +
                 "  Line       : " + fileLine + Environment.NewLine +
                 "  File       : " + filename + Environment.NewLine +
                 "  ToString   : " + e.ToString() + Environment.NewLine +
                 "---";
 
-            Log(ExceptionSeverity, message);
+            Log(_Settings.ExceptionSeverity, message);
         }
 
         /// <summary>
@@ -474,58 +291,50 @@ namespace SyslogLogging
         public virtual void Log(Severity sev, string msg)
         {
             if (String.IsNullOrEmpty(msg)) return;
-            if (sev < MinimumSeverity) return;
+            if (sev < _Settings.MinimumSeverity) return;
 
-            string message = "";
+            string header = "";
             string currMsg = "";
             string remainder = "";
 
-            if (msg.Length > _MaxMessageLength)
+            if (msg.Length > _Settings.MaxMessageLength)
             {
-                currMsg = msg.Substring(0, _MaxMessageLength);
-                remainder = msg.Substring(_MaxMessageLength, (msg.Length - _MaxMessageLength));
+                currMsg = msg.Substring(0, _Settings.MaxMessageLength);
+                remainder = msg.Substring(_Settings.MaxMessageLength, (msg.Length - _Settings.MaxMessageLength));
             }
             else
             {
                 currMsg = msg;
             }
+            
+            header = _Settings.HeaderFormat;
+            if (header.Contains("{ts}")) 
+                header = header.Replace("{ts}", DateTime.Now.ToUniversalTime().ToString(_Settings.TimestampFormat));
+            if (header.Contains("{host}")) 
+                header = header.Replace("{host}", _Hostname);
+            if (header.Contains("{thread}")) 
+                header = header.Replace("{thread}", Thread.CurrentThread.ManagedThreadId.ToString());
+            if (header.Contains("{sev}")) 
+                header = header.Replace("{sev}", sev.ToString());
 
-            if (IncludeUtcTimestamp) message += DateTime.Now.ToUniversalTime().ToString("yyyy-MM-dd HH:mm:ss") + " ";
-            if (IncludeSeverity) message += FormattedSeverity(sev) + " ";
-            if (IncludeHostname) message += _Hostname + " ";
-            if (IncludeThreadId) message += "thr-" + Thread.CurrentThread.ManagedThreadId + " ";
-            if (IndentByStackSize)
+            string message = header + " " + currMsg;
+
+            if (_Settings.EnableConsole)
             {
-                StackTrace st = new StackTrace();
-                int CurrentDepth = st.FrameCount;
-                if (CurrentDepth > _BaseDepth)
-                {
-                    for (int i = 0; i < (CurrentDepth - _BaseDepth); i++)
-                    {
-                        message += " ";
-                    }
-                }
+                SendConsole(sev, message);
             }
 
-            message += currMsg;
-
-            if (ConsoleEnable)
-            {
-                if (!AsyncLogging) SendConsole(sev, message);
-                else Task.Run(() => SendConsole(sev, message));
-            }
-
-            if (!String.IsNullOrEmpty(LogFilename) && FileLogging != FileLoggingMode.Disabled)
+            if (!String.IsNullOrEmpty(_Settings.LogFilename) && _Settings.FileLogging != FileLoggingMode.Disabled)
             {
                 SendFile(sev, message);
             }
 
-            if (_UDP != null)
+            if (_Servers != null && _Servers.Count > 0)
             {
-                if (!AsyncLogging) SendUdp(message);
-                else Task.Run(() => SendUdp(message));
+                List<SyslogServer> servers = new List<SyslogServer>(_Servers);
+                SendServers(servers, message);
             }
-
+             
             if (!String.IsNullOrEmpty(remainder))
             {
                 Log(sev, remainder);
@@ -549,77 +358,52 @@ namespace SyslogLogging
 
             if (disposing)
             {
-                if (_UDP != null)
-                {
-                    try
-                    {
-                        _UDP.Close();
-                    }
-                    catch (Exception)
-                    {
-
-                    }
-                }
+                _TokenSource.Cancel();
             }
 
             _Disposed = true;
         }
 
-        private bool ConsoleExists()
-        {
-            try
-            {
-                bool test1 = Environment.UserInteractive;
-                bool test2 = Console.WindowHeight > 0;
-                return test1 && test2;
-            }
-            catch (Exception)
-            {
-                return false;
-            }
-        }
-
         private void SendConsole(Severity sev, string msg)
         {
             if (String.IsNullOrEmpty(msg)) return;
-            if (!_ConsoleEnable) return;
-
-            if (EnableColors)
+            if (!_Settings.EnableConsole) return;
+            if (_Settings.EnableColors)
             {
                 ConsoleColor prevForeground = Console.ForegroundColor;
                 ConsoleColor prevBackground = Console.BackgroundColor;
 
-                if (Colors != null)
+                if (_Settings.Colors != null)
                 {
                     switch (sev)
                     {
                         case Severity.Debug:
-                            Console.ForegroundColor = Colors.Debug.Foreground;
-                            Console.BackgroundColor = Colors.Debug.Background;
+                            Console.ForegroundColor = _Settings.Colors.Debug.Foreground;
+                            Console.BackgroundColor = _Settings.Colors.Debug.Background;
                             break;
                         case Severity.Info:
-                            Console.ForegroundColor = Colors.Info.Foreground;
-                            Console.BackgroundColor = Colors.Info.Background;
+                            Console.ForegroundColor = _Settings.Colors.Info.Foreground;
+                            Console.BackgroundColor = _Settings.Colors.Info.Background;
                             break;
                         case Severity.Warn:
-                            Console.ForegroundColor = Colors.Warn.Foreground;
-                            Console.BackgroundColor = Colors.Warn.Background;
+                            Console.ForegroundColor = _Settings.Colors.Warn.Foreground;
+                            Console.BackgroundColor = _Settings.Colors.Warn.Background;
                             break;
                         case Severity.Error:
-                            Console.ForegroundColor = Colors.Error.Foreground;
-                            Console.BackgroundColor = Colors.Error.Background;
+                            Console.ForegroundColor = _Settings.Colors.Error.Foreground;
+                            Console.BackgroundColor = _Settings.Colors.Error.Background;
                             break;
                         case Severity.Alert:
-                            Console.ForegroundColor = Colors.Alert.Foreground;
-                            Console.BackgroundColor = Colors.Alert.Background;
+                            Console.ForegroundColor = _Settings.Colors.Alert.Foreground;
+                            Console.BackgroundColor = _Settings.Colors.Alert.Background;
                             break;
                         case Severity.Critical:
-                            Console.ForegroundColor = Colors.Critical.Foreground;
-                            Console.BackgroundColor = Colors.Critical.Background;
+                            Console.ForegroundColor = _Settings.Colors.Critical.Foreground;
+                            Console.BackgroundColor = _Settings.Colors.Critical.Background;
                             break;
                         case Severity.Emergency:
-                            Console.ForegroundColor = Colors.Emergency.Foreground;
-                            Console.BackgroundColor = Colors.Emergency.Background;
+                            Console.ForegroundColor = _Settings.Colors.Emergency.Foreground;
+                            Console.BackgroundColor = _Settings.Colors.Emergency.Background;
                             break;
                     }
                 }
@@ -638,7 +422,7 @@ namespace SyslogLogging
         {
             if (String.IsNullOrEmpty(msg)) return;
 
-            switch (FileLogging)
+            switch (_Settings.FileLogging)
             {
                 case FileLoggingMode.Disabled:
                     return;
@@ -646,12 +430,12 @@ namespace SyslogLogging
                 case FileLoggingMode.SingleLogFile:
                     lock (_FileLock)
                     {
-                        File.AppendAllText(LogFilename, msg + Environment.NewLine);
+                        File.AppendAllText(_Settings.LogFilename, msg + Environment.NewLine);
                     }
                     return;
 
                 case FileLoggingMode.FileWithDate:
-                    string filename = LogFilename + "." + DateTime.Now.ToString("yyyyMMdd");
+                    string filename = _Settings.LogFilename + "." + DateTime.Now.ToString("yyyyMMdd");
                     lock (_FileLock)
                     {
                         File.AppendAllText(filename, msg + Environment.NewLine);
@@ -660,70 +444,30 @@ namespace SyslogLogging
             }
         }
 
-        private void SendUdp(string msg)
+        private void SendServers(List<SyslogServer> servers, string msg)
         {
             if (String.IsNullOrEmpty(msg)) return;
-            
-            lock (_SendLock)
-            {
-                if (_UDP != null)
-                {
-                    byte[] data = Encoding.UTF8.GetBytes(msg);
+            byte[] data = Encoding.UTF8.GetBytes(msg);
 
-                    try
+            foreach (SyslogServer server in servers)
+            {
+                Task.Run(() =>
+                {
+                    lock (server.SendLock)
                     {
-                        _UDP.Send(data, data.Length);
+                        try
+                        {
+                            server.Udp.Send(data, data.Length);
+                        }
+                        catch (Exception)
+                        {
+
+                        }
                     }
-                    catch (Exception)
-                    { 
-                    }
-                }
+                }, _Token).ConfigureAwait(false);
             } 
         }
-
-        private string StackToString()
-        {
-            string ret = "";
-
-            StackTrace t = new StackTrace();
-            for (int i = 2; i < t.FrameCount; i++)
-            {
-                if (i == 2)
-                {
-                    ret += t.GetFrame(i).GetMethod().Name;
-                }
-                else
-                {
-                    ret += " <= " + t.GetFrame(i).GetMethod().Name;
-                }
-            }
-
-            return ret;
-        }
          
-        private string FormattedSeverity(Severity sev)
-        {
-            switch (sev)
-            {
-                case Severity.Debug:
-                    return "[Debug    ]";
-                case Severity.Info:
-                    return "[Info     ]";
-                case Severity.Warn:
-                    return "[Warn     ]";
-                case Severity.Error:
-                    return "[Error    ]";
-                case Severity.Alert:
-                    return "[Alert    ]";
-                case Severity.Critical:
-                    return "[Critical ]";
-                case Severity.Emergency:
-                    return "[Emergency]";
-                default:
-                    throw new ArgumentException("Unknown severity: " + sev.ToString() + ".");
-            }
-        }
-          
         #endregion
     }
 }